Using u32 for alignment positions
As no negative values are valid alignment positions, we
can switch from i64 to u32 maintaining the same possible
range of position values. u32::MAX should more than suffice;
as it is 4,294,967,295.

+++ src/hit.rs
@@ -9,8 +9,8 @@ use std::collections::HashSet;
 #[derive(PartialEq, Eq, Debug, Clone, Default)]
 pub struct Hit {
     pub id: String,
-    pub qstart: i64,
-    pub qend: i64,
+    pub qstart: u32,
+    pub qend: u32,
     pub bitscore: F64,
     pub overlap_with_query: F64,
     pub description: String,
